Add image to `Game` with carrierwave
Games are illustrated to visually tell them apart. We use carrierwave
with minimagick, because it's often used at monterail. Images and
thumbnails are resized to fit, so that they will be easier to
fit on page.
Testing leaves garbage images, so the spec_helper and uploader are
configured to delete garbage.