Initialize 4 models with factories
Models have with basic associations and fields. There is a join table
`UserTeams`, because teams will have many users and teams will have many users.
Matches have two foreign keys that point to teams, so that teams can be
reused in different matches. Rspec is configured to work with FactoryGirl.
Factories are not written yet.